move to internalTextField
