--- cypress/integration/accessibility_TypeaheadV2_spec.js
@@ -1,10 +1,0 @@
-describe('TypeaheadV2 Accessibility check', () => {
-  beforeEach(() => {
-    cy.visit('/TypeaheadV2');
-    cy.injectAxe();
-  });
-
-  it('Tests accessibility on the TypeaheadV2 page', () => {
-    cy.checkA11y();
-  });
-});
