+++ docs/src/components/sidebarIndex.js
@@ -92,7 +92,6 @@ const sidebarIndex: Array<sidebarIndexType> = [
       'TextField',
       'Tooltip',
       'Typeahead',
-      'TypeaheadV2',
     ],
   },
   {
