+++ packages/gestalt/src/InternalTextField.css.flow
@@ -1,0 +1,10 @@
+// @flow strict
+/* This file is automatically generated by css-modules-flow-types */
+declare module.exports: {|
+  +'actionButton': string,
+  +'actionButtonContainer': string,
+  +'textField': string,
+  +'textFieldSpacer': string,
+  +'textFieldWrapper': string,
+  +'unstyledTextField': string,
+|};
