+++ packages/gestalt/src/Typeahead.jsdom.test.js
@@ -6,20 +6,20 @@ import Tag from './Tag.js';
 import Typeahead from './Typeahead.js';
 
 const TOTAL_OPTIONS = 10;
-
-const FAKE_OPTIONS = Array.from(Array(TOTAL_OPTIONS).keys()).map((item) => ({
-  value: `value-${item}`,
-  label: `label-${item}`,
+const FAKE_OPTIONS = new Array(TOTAL_OPTIONS).fill(0).map((item, index) => ({
+  value: `value-${index + 1}`,
+  label: `label-${index + 1}`,
 }));
 
 describe('Typeahead', () => {
   const onBlurMock = jest.fn();
   const onChangeMock = jest.fn();
   const onSelectMock = jest.fn();
-
   const Component = (
     <Typeahead
-      id="Typeahead"
+      clearOptionsLabel="Clear options"
+      showOptionsLabel="Show popup"
+      id="test"
       noResultText="No Result"
       options={FAKE_OPTIONS}
       placeholder="Select a Label"
@@ -32,13 +32,12 @@ describe('Typeahead', () => {
 
   it('renders Typeahead normal', () => {
     const tree = create(Component).toJSON();
-
     expect(tree).toMatchSnapshot();
   });
 
   it('shows menu with data on focus', () => {
     render(Component);
-    const textField = screen.getByRole('textbox', { id: 'Typeahead' });
+    const textField = screen.getByRole('combobox', { id: 'typeahead-test' });
     fireEvent.click(textField);
     const resultsContainer = screen.getAllByText(/label/i);
     expect(resultsContainer.length).toBe(TOTAL_OPTIONS);
@@ -46,65 +45,53 @@ describe('Typeahead', () => {
 
   it('clears menu on blur', () => {
     render(Component);
-    const textField = screen.getByRole('textbox', { id: 'Typeahead' });
+    const textField = screen.getByRole('combobox', { id: 'typeahead-test' });
     textField.click();
     textField.focus();
     const resultsContainer = screen.getAllByText(/label/i);
     expect(resultsContainer.length).toBe(TOTAL_OPTIONS);
-
     act(() => {
       textField.blur();
     });
     expect(onBlurMock).toHaveBeenCalled();
-
     expect(onBlurMock.mock.calls.length).toBe(1);
   });
 
   it('filters menu on search', () => {
     render(Component);
-    const textField = screen.getByRole('textbox', { id: 'Typeahead' });
+    const textField = screen.getByRole('combobox', { id: 'typeahead-test' });
     textField.click();
-
-    fireEvent.change(textField, { target: { value: 'label-3' } });
-
+    fireEvent.change(textField, { target: { value: 'label-1' } });
     const resultsContainer = screen.getAllByText(/label/i);
-    expect(resultsContainer.length).toBe(1);
+    expect(resultsContainer.length).toBe(2);
   });
 
   it('shows no results when no options', () => {
     render(Component);
-    const textField = screen.getByRole('textbox', { id: 'Typeahead' });
+    const textField = screen.getByRole('combobox', { id: 'typeahead-test' });
     textField.focus();
-
-    fireEvent.change(textField, { target: { value: 'No Result' } });
-
+    fireEvent.change(textField, { target: { value: 'label-20' } });
     const resultsContainer = screen.getByText(/no result/i);
     expect(resultsContainer).toBeInTheDocument();
   });
 
   it('calls onChange when typing', () => {
     render(Component);
-    const textField = screen.getByRole('textbox', { id: 'Typeahead' });
+    const textField = screen.getByRole('combobox', { id: 'typeahead-test' });
     textField.focus();
-
     fireEvent.change(textField, { target: { value: 'label' } });
-
     expect(onChangeMock).toHaveBeenCalled();
   });
 
   it('calls onSelect when option is selected', () => {
     render(Component);
-    const textField = screen.getByRole('textbox', { id: 'Typeahead' });
+    const textField = screen.getByRole('combobox', { id: 'typeahead-test' });
     textField.focus();
-
     fireEvent.change(textField, { target: { value: 'label-6' } });
-
     const selectOption = screen.getByText(/label-6/i);
-
     act(() => {
       selectOption.click();
     });
-
     const selectedOption = FAKE_OPTIONS.find((option) => option.label.includes('6'));
     expect(onSelectMock).toHaveBeenCalledWith(expect.objectContaining({ item: selectedOption }));
     expect(onSelectMock.mock.calls.length).toBe(1);
@@ -114,21 +101,27 @@ describe('Typeahead', () => {
     const ref = createRef();
     render(
       <Typeahead
+        clearOptionsLabel="Clear options"
+        showOptionsLabel="Show popup"
         noResultText="No Result"
         label="Ref Example"
-        value="test"
+        value={FAKE_OPTIONS[0].value}
         id="test"
-        options={[{ value: 'test', label: 'test' }]}
+        options={FAKE_OPTIONS}
         ref={ref}
       />,
     );
     expect(ref.current instanceof HTMLInputElement).toEqual(true);
-    expect(ref.current?.value).toEqual('test');
+    expect(ref.current instanceof HTMLInputElement && ref.current.value).toEqual(
+      FAKE_OPTIONS[0].label,
+    );
   });
 
   it('renders tags when supplied', () => {
     const tree = create(
       <Typeahead
+        clearOptionsLabel="Clear options"
+        showOptionsLabel="Show popup"
         noResultText="No Result"
         label="Tag Example"
         value="test"
