--- packages/gestalt/src/TypeaheadInputField.css
@@ -1,62 +1,0 @@
-.icon {
-  composes: marginEnd2 paddingX1 paddingY1 from "./boxWhitespace.css";
-  composes: noBorder from "./Borders.css";
-  composes: pointer from "./Cursor.css";
-  composes: transparentBg from "./Colors.css";
-  outline: none;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-}
-
-.inputRadius {
-  composes: rounding4 from './Borders.css'
-}
-
-.unstyledInput, .inputSpacer {
-  composes: unstyled from "./FormElement.css";
-  composes: Text fontSize3 from "./Text.css";
-  margin-bottom: 4px;
-  padding: 4px 8px;
-}
-
-.unstyledInput {
-  composes: absolute left0 top0 right0 bottom0 from "./Layout.css";
-  width: 100%;
-}
-
-.unstyledInput:focus {
-  outline: 0;
-}
-
-.inputSpacer {
-  visibility: hidden;
-  white-space: nowrap;
-}
-
-.inputWrapper {
-  display: flex;
-  flex-wrap: wrap;
-}
-
-html:not([dir="rtl"]) .inputActive {
-  padding: 8px 32px 8px 16px;
-}
-
-html[dir="rtl"] .inputActive {
-  padding: 8px 16px 8px 32px;
-}
-
-html:not([dir="rtl"]) .icon {
-  right: 0;
-}
-
-html[dir="rtl"] .icon {
-  left: 0;
-}
-
-html:not([dir="rtl"]) .inputWrapper,
-html[dir="rtl"] .inputWrapper {
-  /* Increase specificity for the padding to override SearchField defaults */
-  padding-bottom: 4px;
-}
