--- packages/gestalt/src/TypeaheadInputField.css.flow
@@ -1,10 +1,0 @@
-// @flow strict
-/* This file is automatically generated by css-modules-flow-types */
-declare module.exports: {|
-  +'icon': string,
-  +'inputActive': string,
-  +'inputRadius': string,
-  +'inputSpacer': string,
-  +'inputWrapper': string,
-  +'unstyledInput': string,
-|};
