--- packages/gestalt/src/TypeaheadV2.css
@@ -1,3 +1,0 @@
-.customClass {
-  background: rgba(0, 0, 0, 0.061);
-}
