--- packages/gestalt/src/TypeaheadV2.css.flow
@@ -1,5 +1,0 @@
-// @flow strict
-/* This file is automatically generated by css-modules-flow-types */
-declare module.exports: {|
-  +'customClass': string,
-|};
