--- packages/gestalt/src/TypeaheadV2.flowtest.js
@@ -1,7 +1,0 @@
-// @flow strict
-import TypeaheadV2 from './TypeaheadV2.js';
-
-// const Valid = <TypeaheadV2 />;
-
-// $FlowExpectedError[prop-missing]
-const InvalidProps = <TypeaheadV2 nonexisting={33} />;
