--- packages/gestalt/src/TypeaheadV2.test.js
@@ -1,15 +1,0 @@
-// @flow strict
-// import { create } from 'react-test-renderer';
-// import TypeaheadV2 from './TypeaheadV2.js';
-
-describe('TypeaheadV2', () => {
-  // it('renders', () => {
-  //   const tree = create(<TypeaheadV2 name="" />).toJSON();
-  //   expect(tree).toMatchSnapshot();
-  // });
-
-  // it('renders an accessibility label', () => {
-  //   const tree = create(<TypeaheadV2 accessibilityLabel="Test Accessibility Label" />).toJSON();
-  //   expect(tree).toMatchSnapshot();
-  // });
-});
