+++ packages/gestalt/src/index.js
@@ -1,5 +1,3 @@
-
-
 // @flow strict
 import ActivationCard from './ActivationCard.js';
 import Avatar from './Avatar.js';
@@ -62,7 +60,6 @@ import TextField from './TextField.js';
 import Toast from './Toast.js';
 import Tooltip from './Tooltip.js';
 import Typeahead from './Typeahead.js';
-import TypeaheadV2 from './TypeaheadV2.js';
 import Upsell from './Upsell.js';
 import Video from './Video.js';
 import useFocusVisible from './useFocusVisible.js';
@@ -134,7 +131,6 @@ export {
   Toast,
   Tooltip,
   Typeahead,
-  TypeaheadV2,
   Upsell,
   Video,
   useColorScheme,
